feat(apps/interface): add light mode icon
